+++ origin-js/package.json
@@ -12,8 +12,8 @@
     "format:solidity": "solium -d contracts/contracts/ --fix",
     "install:dev": "npm install && npm link",
     "prepare": "npm run build",
-    "start": "node scripts/build.js serve",
-    "start:no-ganache": "node scripts/build.js serve no-ganache",
+    "start": "node scripts/serve.js serve",
+    "start:no-ganache": "node scripts/serve.js no-ganache",
     "test": "node scripts/test.js",
     "test:truffle": "node scripts/test-truffle.js",
     "test:js": "node scripts/test-js.js",
